requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
tenacity>=8.2.3
pytz>=2023.3
pandas>=2.0.0
